feat: added trash for pages

+++ services/pages.js
@@ -4,12 +4,37 @@ import BadRequestError from '../exceptions/badRequestError.js'
 import ForbiddenError from '../exceptions/forbiddenError.js'
 
 class PagesService {
-  async updatePage(pageId, newPage, userId) {
-    const foundPage = await this.getPage(pageId)
+  async updatePage(pageId, userId, newPage, foundPage = null) {
+    if(!foundPage) {
+      foundPage = await this.getPage(pageId)
+    }
     if(foundPage.author !== userId) throw new ForbiddenError()
     await Page.updateOne({ _id: pageId }, newPage)
 
     return { success: true }
+  }
+
+  async moveToTrash(pageId, userId) {
+    const foundPage = await this.getPage(pageId)
+    if(foundPage.deleted) throw new BadRequestError('Page is already deleted')
+
+    return await this.updatePage(pageId, userId, { deleted: true }, foundPage)
+  }
+
+  async getTrash(userId) {
+    return await Page.find({ author: userId, deleted: true })
+  }
+
+  async deleteAllFromTrash(userId) {
+    await Page.deleteMany({ author: userId, deleted: true })
+    return { success: true }
+  }
+
+  async restorePageFromTrash(pageId, userId) {
+    const foundPage = await this.getPage(pageId)
+    if(!foundPage.deleted) throw new BadRequestError('Page isn\'t deleted')
+
+    return await this.updatePage(pageId, userId, { deleted: false }, foundPage)
   }
 
   async createPage(page) {
@@ -18,6 +43,7 @@ class PagesService {
 
   async deletePage(pageId, userId) {
     const foundPage = await this.getPage(pageId)
+    if(!foundPage.deleted) throw new BadRequestError('Page isn\'t deleted')
     if(foundPage.author !== userId) throw new ForbiddenError()
     await Page.deleteOne({ pageId })
